Fix plumbing error

+++ coral-gulp/tasks/plumb.js
@@ -13,7 +13,7 @@
 const plumber = require('gulp-plumber');
 
 function handleError(err) {
-  global.watchServer.kill();
+  global.watchServer && global.watchServer.kill();
   console.error(err.toString());
   process.exit();
 };
